add interface and additional records(namespace) method.
